Completed HW_5

+++ beauty-salon/src/main/java/com/epam/spring/dto/MasterDto.java
@@ -17,6 +17,7 @@ import java.util.List;
 public class MasterDto extends UserDto {
   @Min(value = 0)
   @Max(value = 5)
-  private Double rating;
+  private Float rating;
+
   private List<Category> categoryList;
 }
